fix: unsubmitted changes check on submit (#1252)
**Context:**
the `detectUnsubmittedChanges` check as implemented had 2 problems:
1. It would fail if the remote-tracking branch wasn't set
2. It relied on the remote-tracking branch instead of the actual remote (meaning it could be out of date if someone else submitted to the branch and we didn't fetch)

**Changes In This Pull Request:**
Use the actual state of remote to do this check.  We already did this for a different check on submit (the check that base of `upstack submit` is up-to-date), so we reused that git call and modified the metaCache callsite to be more generic.

**Test Plan:**
Ran submit locally in different cases.