feat: allow editing PR title and description when updating a PR (#1316)
**Context:**
The CLI currently only allows setting the title and description of a PR when it's first created, and the -e edit flag doesn't work during an update which is a bit surprising.

**Changes In This Pull Request:**
Change the CLI to allow updating the PR title and body when the edit flag is passed in.

**Test Plan:**
Updated this PR with the dev version of the CLI