fix: don't require staged changes on amends (#1320)
**Context:**

https://linear.app/screenplay/issue/GT-6379/communitybug-gt-ca-p-fails-bc-of-no-staged-changes

we're throwing if there are no staged changes when running an amend with no edit. however, git lets you amend with no changes to do things like change the message. also, when you do a patch, the changes are not staged yet and become staged as part of the patch operation so this fixes that also.

**Changes In This Pull Request:**

remove the check for staged changes in commit amend.

**Test Plan:**

added test and ran locally